no thing here

+++ Backend/routes/categoryRouter.js
@@ -1,5 +1,6 @@
 import express from "express";
 import Category from "../Model/categoryModel.js";
+import Post from "../Model/postModel.js";
 import expressAsyncHandler from "express-async-handler";
 
 const categoryRouter = express.Router();
@@ -26,15 +27,33 @@ categoryRouter.post(
     });
   })
 );
+// categoryRouter.delete(
+//   "/:id",
+//   expressAsyncHandler(async (req, res) => {
+//     const category = await Category.findById(req.params.id);
+//     if (category) {
+//       await category.deleteOne();
+//       res.send({ message: "Category Deleted" });
+//     } else {
+//       res.status(404).send({ message: "Category Not Found" });
+//     }
+//   })
+// );
 categoryRouter.delete(
   "/:id",
   expressAsyncHandler(async (req, res) => {
-    const category = await Category.findById(req.params.id);
-    if (category) {
-      await category.deleteOne();
-      res.send({ message: "Category Deleted" });
-    } else {
-      res.status(404).send({ message: "Category Not Found" });
+    const posts = await Post.find({ category: req.params.id });
+    if(posts.length === 0){
+      const category = await Category.findById(req.params.id);
+      if (category) {
+        await category.deleteOne();
+        res.send({ message: "Category Deleted" });
+      } else {
+        res.status(404).send({ message: "Category Not Found" });
+      }
+    }
+    else{
+      res.status(404).send({ message: "This category can be deleted" });
     }
   })
 );
